% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{apply_log_prior}
\alias{apply_log_prior}
\title{Simulated params to evaluable by prior}
\usage{
apply_log_prior(log_prior, theta)
}
\arguments{
\item{theta}{Simulated parameter}
}
\value{
Vector which can be evaluated (log_kappa,v,log_sigma_u,log_sigma_epsilon)
}
\description{
Simulated params to evaluable by prior
}
