% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{calculate_credible_intervals_gaussian}
\alias{calculate_credible_intervals_gaussian}
\title{Calculate credible intervals for Gaussian}
\usage{
calculate_credible_intervals_gaussian(mu, standard_deviation, q)
}
\arguments{
\item{mu}{A vector of length d representing the mean of the Gaussian distribution}

\item{standard_deviation}{A vector of length d representing the standard deviation of the Gaussian distribution}

\item{q}{The significance level}
}
\value{
A matrix of size d x 2 representing the credible intervals for each entry of the Gaussian distribution
}
\description{
Calculates the credible intervals for a Gaussian distribution.
}
\examples{
mu <- c(0, 0)
standard_deviation <- c(1, 1)
q <- 0.05
calculate_credible_intervals_gaussian(mu, standard_deviation, q)
}
