% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{calculate_probabilities}
\alias{calculate_probabilities}
\title{Calculate the probabilities P\link{theta_i <= theta^*_i}}
\usage{
calculate_probabilities(theta_star, theta, w)
}
\arguments{
\item{theta_star}{A vector representing the values theta^*_i.}

\item{theta}{A vector representing the sample.}

\item{w}{A vector of the same length as theta representing the weights.}
}
\value{
A vector of probabilities P\link{theta_i <= theta^*_i}.
}
\description{
Calculates the probabilities P\link{theta_i <= theta^*_i} using the weighted empirical CDF.
}
