% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{check_range}
\alias{check_range}
\title{Helper function to check range of parameters}
\usage{
check_range(x, name, lower, upper)
}
\arguments{
\item{x}{A number}

\item{name}{A string}

\item{lower}{A number}

\item{upper}{A number}
}
\value{
A warning if x is not in the range \link{lower,upper}
}
\description{
Checks if the parameters are in the correct range
}
\examples{
check_range(x = 0.5, name = "x", lower = 0, upper = 1)
}
