% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{complexity}
\alias{complexity}
\title{Complexity}
\usage{
complexity(log_kappa, v)
}
\arguments{
\item{log_kappa}{Logarithm of kappa}

\item{v}{A two dimensional vector that controls the direction and magnitude of anisotropy}
}
\value{
The calculated complexity of the model
}
\description{
Calculates the complexity of the model
}
\examples{
log_kappa <- -0.3
v <- c(1, 2)
complexity(log_kappa, v)
}
