% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{correlation_range}
\alias{correlation_range}
\title{Correlation range}
\usage{
correlation_range(log_kappa)
}
\arguments{
\item{log_kappa}{The logarithm of kappa}
}
\value{
The correlation range
}
\description{
Calculates the correlation range given log(kappa)
}
\examples{
log_kappa <- -1
correlation_range(log_kappa)
}
