% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{fdist}
\alias{fdist}
\title{fdist}
\usage{
fdist(r)
}
\arguments{
\item{r}{A positive number representing the norm of the anisotropy vector}
}
\value{
The calculated component of the distance from the flexible to the base model depending on |v|
}
\description{
Calculates the component of the distance from the flexible to the base model depending on |v|
}
\examples{
r <- 0.5
fdist(r)
}
