% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{fdist_prime}
\alias{fdist_prime}
\title{fdist_prime}
\usage{
fdist_prime(r)
}
\arguments{
\item{r}{A positive number representing the norm of the anisotropy vector}
}
\value{
The calculated derivative of the component of the distance from the flexible to the base model depending on |v|
}
\description{
Calculates the derivative of the compenent of the distance from the flexible to the base model depending on |v|
}
\examples{
r <- 0.5
fdist_prime(r)
}
