% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{KL_discrete_log_unnormalized_weights}
\alias{KL_discrete_log_unnormalized_weights}
\title{KL discrete log unnormalized weights}
\usage{
KL_discrete_log_unnormalized_weights(
  log_unnormalized_weights_1,
  log_unnormalized_weights_2
)
}
\arguments{
\item{log_unnormalized_weights_1}{A vector of log unnormalized weights for the first distribution}

\item{log_unnormalized_weights_2}{A vector of log unnormalized weights for the second distribution}
}
\value{
The KL divergence between the two distributions
}
\description{
Calculates the KL divergence between two discrete distributions with the same support using the log unnormalized weights
}
\examples{
log_unnormalized_weights_1 <- log(c(1, 2, 3))
log_unnormalized_weights_2 <- log(c(3, 2, 1))
KL_discrete_log_unnormalized_weights(log_unnormalized_weights_1, log_unnormalized_weights_2)
}
