% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{lambda1_quantile}
\alias{lambda1_quantile}
\title{Hyperparameter lambda1 for given quantile of the PC prior of |v|}
\usage{
lambda1_quantile(a0, beta = 0.01)
}
\arguments{
\item{a0}{A surprisingly high ratio of anisotropy}

\item{beta}{A quantile in \link{0,1} of the PC prior of |v|}
}
\description{
Calculates  the hyperparameter lambda1 for given quantile beta and large anisotropy a0 of the PC prior of |v|
}
\examples{
beta <- 0.01
a0 <- 10
lambda1 <- lambda1_quantile(r0, beta= 0.01)
}
