% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{lambda_quantile}
\alias{lambda_quantile}
\title{Hyperparameter lambda for given quantile of the PC prior of kappa}
\usage{
lambda_quantile(rho0, lambda1, alpha = 0.01)
}
\arguments{
\item{rho0}{A surprisingly large correlation range}

\item{alpha}{A quantile in (0,1) of the PC prior of kappa}
}
\description{
Calculates  the hyperparameter lambda for given quantile alpha, tolerance kappa0 and lambda1 of the PC prior of kappa
}
\examples{
alpha <- 0.01
kappa0 <- 0.1
lambda1 <- 1
lambda <- lambda_quantile(alpha=alpha, kappa0=kappa0, lambda1=lambda1)
}
