% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{log_beta_density}
\alias{log_beta_density}
\title{Log beta density}
\usage{
log_beta_density(a, b, shape)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{shape}{A shape parameter of the beta distribution}
}
\value{
The function for the log of the beta density on \link{a,b}
}
\description{
Calculates the log of the beta density on \link{a,b} centred at the midpoint with shape parameters shape
}
\examples{
a <- 0
b <- 1
shape <- 1.1
log_beta_density <- log_beta_density(a = a, b = b, shape = shape)
}
