% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{log_pc_prior_quantile}
\alias{log_pc_prior_quantile}
\title{Log PC prior on theta=(log(kappa),v, log(sigma_u), log(sigma_epsilon)) given certain quantiles.}
\usage{
log_pc_prior_quantile(
  sigma_u0,
  sigma_epsilon0,
  a0,
  rho0,
  alpha = 0.01,
  alpha_u = NULL,
  alpha_epsilon = NULL,
  alpha_k = NULL,
  alpha_v = NULL
)
}
\arguments{
\item{sigma_u0}{A surprisingly high variance of field, in (0,infinity)}

\item{sigma_epsilon0}{A surprisingly high variance of noise, in (0,infinity)}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}

\item{rho0}{A surprisingly small correlation range, in (0,infinity)}

\item{alpha}{A quantile in (0,1) for all the parameters. By default, alpha = 0.01}

\item{alpha_u}{A quantile in (0,1) for the variance of the field}

\item{alpha_epsilon}{A quantile in (0,1) for the variance of the noise}

\item{alpha_k}{A quantile in (0,1) for kappa}

\item{alpha_v}{A quantile in (0,1) for anisotropy |v|}
}
\value{
The calculated log of the PC prior on  theta=(log(kappa),v, log(sigma_u), log(sigma_epsilon))
}
\description{
Calculates  the log of the PC prior on the (log(kappa),v, log(sigma_u), log(sigma_epsilon)) given certain quantiles.
}
\examples{
alpha <- 0.05
sigma_u0 <- 10
sigma_epsilon0 <- 1
a0 <- 10
rho0 <- 0.1
log_pc_prior <- log_pc_prior_quantile(sigma_u0 = sigma_u0, sigma_epsilon0 = sigma_epsilon0, a0 = a0, rho0 = rho0, alpha = alpha)
}
