% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{log_pc_prior_theta}
\alias{log_pc_prior_theta}
\title{Log PC Prior calculation for theta=(log(kappa), v, log(sigma_u), log(sigma_epsilon))}
\usage{
log_pc_prior_theta(
  lambda,
  lambda1,
  lambda_epsilon,
  lambda_u,
  log_kappa,
  v,
  log_sigma_u,
  log_sigma_epsilon
)
}
\arguments{
\item{lambda}{A hyperparameter controlling the size of kappa.}

\item{lambda1}{A hyperparameter controlling the size of |v|.}

\item{lambda_epsilon}{A hyperparameter controlling the size of epsilon.}

\item{lambda_u}{A hyperparameter controlling the size of sigma_u.}

\item{log_kappa}{Logarithm of inverse correlation range.}

\item{v}{Vector that controls anisotropy.}

\item{log_sigma_u}{Logarithm of the variance of the field.}

\item{log_sigma_epsilon}{The logarithm of the variance of the additive noise.}
}
\value{
The calculated log PC prior value of theta=(log(kappa), v, log(sigma_u), log(sigma_epsilon)).
}
\description{
Calculates  the log of the PC prior for theta=(log(kappa), v, log(sigma_u), log(sigma_epsilon)) based on given hyperparameters and vectors.
}
\examples{
lambda <- 1
lambda1 <- 1
lambda_epsilon <- 1
lambda_u <- 1
log_kappa <- 0.5
v <- c(1, 2)
log_sigma_u <- 0
log_sigma_epsilon <- 0
log_pc_prior_theta(lambda = lambda, lambda1 = lambda1, lambda_epsilon = lambda_epsilon, lambda_u = lambda_u, log_kappa = log_kappa, v = v, log_sigma_u = log_sigma_u, log_sigma_epsilon = log_sigma_epsilon)
}
