% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{log_prior_aniso_quantile_gaussian}
\alias{log_prior_aniso_quantile_gaussian}
\title{Gaussian prior on anisotropy given quantiles}
\usage{
log_prior_aniso_quantile_gaussian(alpha_k, alpha_v, rho0, a0)
}
\arguments{
\item{alpha_k}{A quantile in (0,1) for kappa}

\item{alpha_v}{A quantile in (0,1) for |v|}

\item{rho0}{A surprisingly small correlation range}

\item{a0}{A surprisingly high ratio of anisotropy}
}
\value{
The calculated log of the gaussian prior on the (log(kappa),v)
}
\description{
Calculates  the log of the gaussian prior on the (log(kappa),v) supposing log(|v|)~N(1,sigma_v^2) and log(kappa)~N(1,sigma_k^2)
such that the anisotropy ratio exp(2|v|) is larger than a0 with probability alpha and the correlation range sqrt{16}/kappa is smaller than rho0 with probability alpha
}
\examples{
alpha_k <- 0.01
alpha_v <- 0.01
a0 <- 10
rho0 <- 1
log_prior_aniso_gaussian <- log_prior_aniso_quantile(alpha_k = alpha_k, beta_v = beta_v, rho0 = rho0, a0 = a0)
}
