% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{log_prior_quantile_gaussian}
\alias{log_prior_quantile_gaussian}
\title{Log prior on anisotropy, noise and variance of field supposing log(|v|)~N(1,sigma_v^2) and log(kappa)~N(1,sigma_k^2)
and with PC priors on noise and variance of field, given certain quantiles.}
\usage{
log_prior_quantile_gaussian(
  alpha_u,
  alpha_epsilon,
  alpha_k,
  alpha_v,
  sigmau0,
  sigmaepsilon0,
  a0,
  rho0
)
}
\arguments{
\item{alpha_u}{A quantile in (0,1) for the variance of the field}

\item{alpha_epsilon}{A quantile in (0,1) for the variance of the noise}

\item{alpha_k}{A quantile in (0,1) for kappa}

\item{alpha_v}{A quantile in (0,1) for |v|}

\item{sigmau0}{A surprisingly high variance of field, in (0,infinity)}

\item{sigmaepsilon0}{A surprisingly high variance of noise, in (0,infinity)}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}

\item{rho0}{A surprisingly small correlation range, in (0,infinity)}
}
\value{
The calculated log of the prior on the (log(kappa),v, log(sigma_u), log(sigma_epsilon))
}
\description{
Calculates  the log of the prior on the (log(kappa),v, log(sigma_u), log(sigma_epsilon))
supposing log(|v|)~N(1,sigma_v^2) and log(kappa)~N(1,sigma_k^2)
and with PC priors on noise and variance of field, given certain quantiles.
such that the anisotropy ratio exp(2|v|) is larger than a0 with probability alpha and the correlation range sqrt{16}/kappa is smaller than rho0 with probability alpha
}
\examples{
alpha_u <- 0.01
alpha_epsilon <- 0.01
alpha_k <- 0.01
alpha_v <- 0.01
sigmau0 <- 10
sigmaepsilon0 <- 1
a0 <- 10
rho0 <- 0.1
log_prior_gaussian <- log_prior_aniso_quantile(alpha_u = alpha_u, alpha_epsilon = alpha_epsilon, alpha_k = alpha_k, alpha_v = alpha_v, sigmau0 = sigmau0, sigmaepsilon0 = sigmaepsilon0, a0 = a0, rho0 = rho0)
}
