% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{log_uniform_density}
\alias{log_uniform_density}
\title{Log uniform density}
\usage{
log_uniform_density(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The function for the log of the uniform density on \link{a,b}
}
\description{
Calculates the log of the uniform density on \link{a,b}
}
\examples{
a <- 0
b <- 1
log_uniform_density <- log_uniform_density(a = a, b = b)
log_uniform_density(0.5)
}
