% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{log_unnormalized_importance_weights_and_integrals}
\alias{log_unnormalized_importance_weights_and_integrals}
\title{Calculate unnormalized log importance weights, mean, variance, and KL divergences}
\usage{
log_unnormalized_importance_weights_and_integrals(
  log_posterior_density,
  mu_Laplace,
  Q_Laplace,
  n_weights,
  q = 0.05,
  true_params
)
}
\arguments{
\item{log_posterior_density}{A function that calculates the unnormalized log posterior of theta}

\item{mu_Laplace}{A 5D vector representing the mean of the Laplace approximation}

\item{Q_Laplace}{A 5x5 matrix representing the precision matrix of the Laplace approximation}

\item{n_weights}{Number of weights to be calculated}

\item{q}{The significance level. By default, set to 0.05.}
}
\value{
A list with three elements: log_unnormalized_weights, log_unnormalized_weights_smoothed and the k-diagnostic for PSIS
}
\description{
Calculates the unnormalized log importance weights, mean, variance, and KL divergences between the importance approximation and the Laplace approximation
}
\examples{
# Defines the log prior density function of theta for PC and non-PC priors
log_pc_prior <- log_pc_prior_quantile(
  sigma_u0 = sigma_u0, sigma_epsilon0 = sigma_epsilon0,
  a0 = a0, rho0 = rho0, alpha = alpha
)
# Defining the log posterior density as a function of the parameters using the log_pc_prior
m_u <- 0
log_posterior_pc <- function(theta) {
  log_kappa <- theta[1]
  v <- theta[2:3]
  log_sigma_u <- theta[4]
  log_sigma_epsilon <- theta[5]
  log_posterior_prior(
    log_prior = log_pc_prior,
    mesh = mesh, log_kappa = log_kappa, v = v,
    log_sigma_epsilon = log_sigma_epsilon, log_sigma_u = log_sigma_u,
    y = y, A = A, m_u = m_u
  )
}
# Calculates the importance weights, mean, variance, and KL divergences
mu_Laplace <- c(0, 0, 0, 0, 0)
Q_Laplace <- diag(5)
log_unnormalized_importance_weights_and_integrals(log_posterior_pc, mu_Laplace, Q_Laplace, n_weights = 1000)
}
