% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{MAPgeneral}
\alias{MAPgeneral}
\title{Calculates  the MAP estimate for linear noisy observation of field with a general prior on (kappa,v).}
\usage{
MAPgeneral(
  logprior_aniso,
  mesh,
  lambda,
  lambda1,
  lambda_epsilon,
  lambda_u,
  y,
  A,
  m_u,
  maxiterations = 300,
  log_sigma_epsilon = NULL
)
}
\arguments{
\item{logprior_aniso}{A function that calculates the log prior of (kappa, v)}

\item{mesh}{The mesh}

\item{lambda}{A hyperparameter controlling the size of kappa.}

\item{lambda1}{A hyperparameter controlling the size of |v|.}

\item{lambda_epsilon}{A hyperparameter controlling the size of sigma_epsilon.}

\item{y}{A vector with length equal to the number of basis elements n representing the observed data.}

\item{A}{Matrix of size nxn representing the transformation A}

\item{m_u}{A vector with length n representing the prior mean m_u}

\item{maxiterations}{Maximum number of iterations for optim, by default 300}

\item{log_sigma_epsilon}{Variance of noise, if NULL it is estimated by the MAP}
}
\value{
The parameters (log_kappa, v, log_sigma_u, log_sigma_epsilon) that maximize the posterior
}
\description{
Calculated by maximizing log posterior using optim. Only stationary parameters accepted.
}
