% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mean_distance_to_MAP_and_std_dev_of_Gaussian_approximation}
\alias{mean_distance_to_MAP_and_std_dev_of_Gaussian_approximation}
\title{Mean distances to MAP and standard deviation of Gaussian approximation around the median}
\usage{
mean_distance_to_MAP_and_std_dev_of_Gaussian_approximation(
  results,
  prior_types
)
}
\arguments{
\item{results}{A list of results from the simulation containing the distances to MAP}

\item{prior_types}{A character vector of prior types}
}
\value{
A list of vectors containing the mean distances to MAP and the standard deviation of the Gaussian approximation around the median for each prior type
}
\description{
Gets the mean distances to MAP and the standard deviation of the Gaussian approximation around the median for each prior type
}
