% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{normalize_log_weights}
\alias{normalize_log_weights}
\title{Normalized importance weights}
\usage{
normalize_log_weights(log_unnormalized_weights)
}
\arguments{
\item{log_unnormalized_weights}{A vector of log unnormalized weights}
}
\value{
A vector of normalized importance weights
}
\description{
Given log unnormalized weights, calculates the normalized importance weights
}
\examples{
log_unnormalized_weights <- log(c(1, 2, 3))
normalize_log_weights(log_unnormalized_weights)
}
