% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{parameter_within_confidence_intervals}
\alias{parameter_within_confidence_intervals}
\title{Parameter within confidence intervals}
\usage{
parameter_within_confidence_intervals(parameter, confidence_intervals)
}
\arguments{
\item{parameter}{A vector of length d representing the parameter}

\item{confidence_intervals}{A matrix of size d x 2 representing the confidence intervals}
}
\value{
A vector of length d representing whether the parameter is within the confidence intervals
}
\description{
Checks if the parameter is within the confidence intervals
}
\examples{
mu <- c(0, 0)
standard_deviation <- c(1, 1)
q <- 0.05
confidence_intervals <- calculate_confidence_intervals_gaussian(mu, standard_deviation, q)
theta <- c(0.5, 0.5)
parameter_within_confidence_intervals(theta, confidence_intervals)

}
