% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_CI_lengths_and_get_mean_lengths}
\alias{plt_CI_lengths_and_get_mean_lengths}
\title{Plot credible interval lengths and get mean lengths}
\usage{
plt_CI_lengths_and_get_mean_lengths(
  results,
  prior_types,
  approximation_types,
  parameter_names,
  path = NULL,
  dpi = 600,
  width = 10,
  height = 10
)
}
\arguments{
\item{results}{A list of results from the simulation containing the credible intervals}

\item{prior_types}{A character vector of prior types}

\item{approximation_types}{A character vector of approximation types}

\item{parameter_names}{A character vector of parameter names}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}
}
\value{
A list of data frames containing the mean lengths for each prior type and approximation type
}
\description{
Plot credible interval lengths and get mean lengths for each prior type and approximation type
}
