% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_complexity_and_get_mean_complexity}
\alias{plt_complexity_and_get_mean_complexity}
\title{Plot complexity and get mean complexity}
\usage{
plt_complexity_and_get_mean_complexity(
  results,
  prior_types,
  approximation_types,
  path = NULL,
  dpi = 600,
  width = 10,
  height = 10
)
}
\arguments{
\item{results}{A list of results from the simulation containing the complexity}

\item{prior_types}{A character vector of prior types}

\item{approximation_types}{A character vector of approximation types}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}
}
\value{
A list of vectors containing the mean complexity for each prior type and approximation type
}
\description{
Plot complexity and get mean complexity for each prior type and approximation type
}
