% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_distances_to_MAP}
\alias{plt_distances_to_MAP}
\title{Plot distances to MAP}
\usage{
plt_distances_to_MAP(
  results,
  prior_types,
  path = NULL,
  dpi = 600,
  width = 10,
  height = 10
)
}
\arguments{
\item{results}{A list of results from the simulation containing the distances to MAP}

\item{prior_types}{A character vector of prior types}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot. By default, it is 600}

\item{width}{An integer of the width of the plot. By default, it is 10}

\item{height}{An integer of the height of the plot. By default, it is 10}
}
\value{
A ggplot object
}
\description{
Plot distances to MAP for each prior type
}
