% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_frequency_true_parameter_in_CI}
\alias{plt_frequency_true_parameter_in_CI}
\title{Plot the frequency of the true parameter being in the credible interval}
\usage{
plt_frequency_true_parameter_in_CI(
  results = results,
  prior_types = prior_types,
  approximation_types = approximation_types,
  parameter_names = parameter_names,
  path = NULL,
  dpi = 600,
  width = 10,
  height = 10
)
}
\arguments{
\item{results}{A list of results from the simulation containing the credible intervals}

\item{prior_types}{A character vector of prior types}

\item{approximation_types}{A character vector of approximation types}

\item{parameter_names}{A character vector of parameter names}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}
}
\description{
Plot the frequency of the true parameter being in the credible interval for each prior type and approximation type
}
