% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_k_diagnostics}
\alias{plt_k_diagnostics}
\title{Plot k diagnostics}
\usage{
plt_k_diagnostics(
  results,
  prior_types,
  path = NULL,
  dpi = 100,
  width = 10,
  height = 10
)
}
\arguments{
\item{results}{A list of results from the simulation containing the k diagnostics}

\item{prior_types}{A character vector of prior types}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}

\item{approximation_types}{A character vector of approximation types}
}
\description{
Plot k diagnostics for each prior type
}
