% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{seconds_to_hours_minutes_seconds}
\alias{seconds_to_hours_minutes_seconds}
\title{Convert seconds to hours, minutes, and seconds}
\usage{
seconds_to_hours_minutes_seconds(seconds)
}
\arguments{
\item{seconds}{A number representing the number of seconds}
}
\value{
A string representing the number of hours, minutes, and seconds
}
\description{
Converts seconds to hours, minutes, and seconds
}
\examples{
seconds <- 100
seconds_to_hours_minutes_seconds(seconds)
seconds <- 10000
seconds_to_hours_minutes_seconds(seconds)
}
