% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sigma2_quantile_v}
\alias{sigma2_quantile_v}
\title{Parameter sigma^2 such that if v1,v2 ~ N(0, sigma^2) the anisotropy ratio exp(|v|) is larger than a0 with probability alpha}
\usage{
sigma2_quantile_v(a0, alpha_v = 0.01)
}
\arguments{
\item{a0}{A surprisingly high ratio of anisotropy}

\item{alpha_v}{A quantile in (0,1). By default 0.01}
}
\value{
The calculated parameter sigma^2
}
\description{
Calculates the parameter sigma^2 such that if v1,v2 are iid N(0, sigma^2), then the anisotropy ratio exp(2|v|) is larger than a0 with probability alpha
}
\examples{
alpha_v <- 0.01
a0 <- 10
sigma2_v <- sigma2_quantile_v(alpha_v = alpha_v, a0 = a0)
}
