% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sigma_quantile}
\alias{sigma_quantile}
\title{This calculates sigma such that if x~N(0,sigma^2) then l>= x0 with probability alpha}
\usage{
sigma_quantile(alpha, x0)
}
\arguments{
\item{alpha}{A quantile in (0,1)}

\item{x0}{A surprisingly high number}
}
\value{
The calculated parameter sigma^2
}
\description{
Calculates  the parameter sigma^2 such that if l~N(0,sigma^2) then x>x0 with probability alpha
}
\examples{
alpha <- 0.01
x0 <- 10
sigma <- sigma_quantile(alpha = alpha, x0 = x0)
}
