% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sigma_quantile_kappa}
\alias{sigma_quantile_kappa}
\title{Parameter sigma such that if log(kappa)~N(1,sigma) the correlation range sqrt{16}/kappa is smaller than rho0 with probability alpha}
\usage{
sigma_quantile_kappa(alpha_k, rho0)
}
\arguments{
\item{alpha_k}{A quantile in (0,1)}

\item{rho0}{A surprisingly small correlation range}
}
\value{
The calculated parameter sigma
}
\description{
Calculates  the parameter sigma such that if log(kappa) is gaussian then the correlation range sqrt{16}/kappa is smaller than rho0 with probability alpha
}
\examples{
alpha_k <- 0.01
rho0 <- 1
sigma_k <- sigma_quantile_kappa(alpha_k = alpha_k, rho0 = rho0)
}
