% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sigma_quantile_v}
\alias{sigma_quantile_v}
\title{Parameter sigma such that if log(|v|)~N(1,sigma^2) the anisotropy ratio exp(2|v|) is larger than a0 with probability alpha}
\usage{
sigma_quantile_v(alpha_v, a0)
}
\arguments{
\item{alpha_v}{A quantile in (0,1)}

\item{a0}{A surprisingly high ratio of anisotropy}
}
\value{
The calculated parameter sigma^2
}
\description{
Calculates  the parameter sigma^2 such that if log(|v|) is gaussian then the anisotropy ratio exp(2|v|) is larger than a0 with probability alpha
}
\examples{
alpha_v <- 0.01
a0 <- 10
sigma_v <- sigma_quantile_v(alpha_v = alpha_v, a0 = a0)
}
