% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sim_not_pc}
\alias{sim_not_pc}
\title{Sim not_pc: log(kappa)~N(1,sigma_k^2) and v1,v2~N(0,sigma_v^2)
and with PC priors on noise and variance of field, given certain quantiles.}
\usage{
sim_not_pc(
  sigma_u0,
  sigma_epsilon0,
  a0,
  rho0,
  alpha = 0.01,
  alpha_u = NULL,
  alpha_epsilon = NULL,
  alpha_k = NULL,
  alpha_v = NULL,
  m = 1
)
}
\arguments{
\item{sigma_u0}{A surprisingly high variance of field, in (0,infinity)}

\item{sigma_epsilon0}{A surprisingly high variance of noise, in (0,infinity)}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}

\item{rho0}{A surprisingly small correlation range, in (0,infinity)}

\item{alpha}{A quantile in (0,1) for all the parameters. By default, alpha = 0.01}

\item{alpha_u}{A quantile in (0,1) for the variance of the field. If NULL, alpha_u=alpha}

\item{alpha_epsilon}{A quantile in (0,1) for the variance of the noise. If NULL, alpha_epsilon=alpha}

\item{alpha_k}{A quantile in (0,1) for kappa. If NULL, alpha_k=alpha}

\item{alpha_v}{A quantile in (0,1) for anisotropy |v|. If NULL, alpha_v=alpha}

\item{m}{Number of samples, by default 1.}
}
\value{
The simulated values of log(kappa), v, log(sigma_u), log(sigma_epsilon)
}
\description{
CSimulates from (log(kappa),v, log(sigma_u), log(sigma_epsilon))
supposing log(|v|)~N(1,sigma_v^2) and kappa~Exp(lambda_k)
and with PC priors on noise and variance of field, given certain quantiles.
such that the anisotropy ratio exp(|v|) is larger than a0 with probability alpha and the correlation range sqrt{8}/kappa is smaller than rho0 with probability alpha
}
\examples{
alpha_u <- 0.01
alpha_epsilon <- 0.01
alpha_k <- 0.01
alpha_v <- 0.01
sigma_u0 <- 10
sigma_epsilon0 <- 1
a0 <- 10
rho0 <- 0.1
m <- 10
result <- sim_not_pc(alpha_u = alpha_u, alpha_epsilon = alpha_epsilon, alpha_k = alpha_k, alpha_v = alpha_v, sigma_u0 = sigma_u0, sigma_epsilon0 = sigma_epsilon0, a0 = a0, rho0 = rho0, m = m)
}
