% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sim_theta_beta}
\alias{sim_theta_beta}
\title{Simulation of theta given quantiles for the beta prior}
\usage{
sim_theta_beta(
  sigma_u0,
  sigma_epsilon0,
  a0,
  rho0,
  L,
  shape,
  width_support_factor = 2,
  alpha = 0.01,
  alpha_u = NULL,
  alpha_epsilon = NULL,
  m = 1
)
}
\arguments{
\item{sigma_u0}{A surprisingly high variance of field, in (0,infinity)}

\item{sigma_epsilon0}{A surprisingly high variance of noise, in (0,infinity)}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity), anisotropy ratio is not allowed to go below a0/2 or above 2a0}

\item{rho0}{A surprisingly small correlation range, correlation is not allowed to go below rho0/2}

\item{L}{A surprisingly large correlation range (e.g. size of domain), correlation is not allowed to go above L}

\item{shape}{A shape parameter of the beta distribution}

\item{width_support_factor}{A factor to multiply the width of the support of the uniform distribution. By default 2}

\item{alpha}{A quantile in (0,1) for all the parameters. By default, alpha = 0.01}

\item{alpha_u}{A quantile in (0,1) for the variance of the field. If NULL, alpha_u=alpha}

\item{alpha_epsilon}{A quantile in (0,1) for the variance of the noise. If NULL, alpha_epsilon=alpha}

\item{m}{Number of samples, by default 1.}
}
\value{
A list with four elements: log_kappa, v, log_sigma_u, log_sigma_epsilon
}
\description{
Simulates theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon)) from the beta prior. That is:
correlation range rho = sqrt(8)/kappa ~ beta(shape,shape) with support (rho0/2,L),
v1,v2 ~beta(shape,shape)with support (log(a0/w)/sqrt(2),log(wa0)/sqrt(2)), and with PC priors on noise and variance of field, given certain quantiles.
Both shape parameters of the beta distribution are set to shape
}
\examples{
alpha_u <- 0.01
alpha_epsilon <- 0.01
sigma_u0 <- 10
sigma_epsilon0 <- 1
a0 <- 10
rho0 <- 0.1
L <- 100
width_support_factor <- 2
shape <- 1.1
m <- 10
result <- sim_theta_beta(alpha_u = alpha_u, alpha_epsilon = alpha_epsilon, sigma_u0 = sigma_u0, sigma_epsilon0 = sigma_epsilon0, a0 = a0, rho0 = rho0, L = L, shape = shape, width_support_factor = width_support_factor, m = m)
}
