% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{sim_theta_pc}
\alias{sim_theta_pc}
\title{Simulation of theta under PC priors}
\usage{
sim_theta_pc(lambda, lambda1, lambda_u, lambda_epsilon, m = 1)
}
\arguments{
\item{lambda}{A hyperparameter controlling the size of kappa.}

\item{lambda1}{A hyperparameter controlling the size of |v|.}

\item{lambda_u}{A hyperparameter controlling the size of sigma_u.}

\item{lambda_epsilon}{A hyperparameter controlling the size of sigma_epsilon.}

\item{m}{Number of samples, by default 1.}
}
\value{
A list with four elements: log_kappa, v, log_sigma_u, log_sigma_epsilon
}
\description{
Simulates theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon)) from the PC prior
}
\examples{
lambda <- 1
lambda1 <- 1
lambda_epsilon <- 1
lambda_u <- 1
m <- 10
result <- sim_theta_pc(lambda = lambda, lambda1 = lambda1, lambda_epsilon = lambda_epsilon, lambda_u = lambda_u, m = m)
}
