% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sim_theta_pc_quantile}
\alias{sim_theta_pc_quantile}
\title{Simulation of theta given quantiles of the PC prior}
\usage{
sim_theta_pc_quantile(
  sigmau0,
  sigmaepsilon0,
  a0,
  rho0,
  alpha = 0.01,
  alpha_u = NULL,
  alpha_epsilon = NULL,
  alpha_k = NULL,
  alpha_v = NULL,
  m = 1
)
}
\arguments{
\item{sigmau0}{A surprisingly high variance of field, in (0,infinity)}

\item{sigmaepsilon0}{A surprisingly high variance of noise, in (0,infinity)}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}

\item{rho0}{A surprisingly small correlation range, in (0,infinity)}

\item{alpha}{A quantile in (0,1) for all the parameters. By default, alpha = 0.01}

\item{alpha_u}{A quantile in (0,1) for the variance of the field}

\item{alpha_epsilon}{A quantile in (0,1) for the variance of the noise}

\item{alpha_k}{A quantile in (0,1) for kappa}

\item{alpha_v}{A quantile in (0,1) for anisotropy |v|}

\item{m}{Number of samples, by default 1.}
}
\value{
A list with four elements: log_kappa, v, log_sigma_u, log_sigma_epsilon
}
\description{
Simulates theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon)) from the PC prior given quantiles of the PC prior
}
\examples{
alpha <- 0.05
sigmau0 <- 10
sigmaepsilon0 <- 1
a0 <- 10
rho0 <- 0.1
m <- 10
result <- sim_theta_pc_quantile(alpha = alpha, sigmau0 = sigmau0, sigmaepsilon0 = sigmaepsilon0, a0 = a0, rho0 = rho0, m = m)
}
