% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{sim_theta_uniform}
\alias{sim_theta_uniform}
\title{Simulation of theta given quantiles for the uniform prior}
\usage{
sim_theta_uniform(
  sigma_u0,
  sigma_epsilon0,
  a0,
  rho0,
  L,
  alpha = 0.01,
  alpha_u = NULL,
  alpha_epsilon = NULL
)
}
\arguments{
\item{sigma_u0}{A surprisingly high variance of field, in (0,infinity)}

\item{sigma_epsilon0}{A surprisingly high variance of noise, in (0,infinity)}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}

\item{rho0}{A surprisingly small correlation range, in (0,infinity)}

\item{L}{A surprisingly large correlation range (e.g. size of domain), in (0,infinity)}

\item{alpha}{A quantile in (0,1) for all the parameters. By default, alpha = 0.01}

\item{alpha_u}{A quantile in (0,1) for the variance of the field. If NULL, alpha_u=alpha}

\item{alpha_epsilon}{A quantile in (0,1) for the variance of the noise. If NULL, alpha_epsilon=alpha}
}
\value{
A list with four elements: log_kappa, v, log_sigma_u, log_sigma_epsilon
}
\description{
Simulates theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon)) from the uniform prior given quantiles for the uniform prior
}
\examples{
alpha_u <- 0.01
alpha_epsilon <- 0.01
sigma_u0 <- 10
sigma_epsilon0 <- 1
a0 <- 10
rho0 <- 0.1
L <- 100
m <- 10
result <- sim_theta_uniform(alpha_u = alpha_u, alpha_epsilon = alpha_epsilon, sigma_u0 = sigma_u0, sigma_epsilon0 = sigma_epsilon0, a0 = a0, rho0 = rho0, L = L)
}
