% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{calculate_credible_intervals_importance}
\alias{calculate_credible_intervals_importance}
\title{Calculate credible intervals for each entry of a weighted sample}
\usage{
calculate_credible_intervals_importance(theta, w, q)
}
\arguments{
\item{theta}{A matrix of size n x d representing the weighted sample}

\item{w}{A vector of length n representing the weights of the unnormalized weighted sample.}

\item{q}{The significance level}
}
\value{
A matrix of size d x 2 representing the credible intervals for each entry of the weighted sample
}
\description{
Calculates the credible intervals for each entry of a weighted sample.
}
