% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{lambda_variance_quantile}
\alias{lambda_variance_quantile}
\title{Hyperparameter for variance of gaussian field given quantiles}
\usage{
lambda_variance_quantile(sigma0, alpha_sigma = 0.01)
}
\arguments{
\item{sigma0}{A surprisingly high number}

\item{alpha_sigma}{A quantile in (0,1). By default 0.01}
}
\value{
The calculated hyperparameter lambda
}
\description{
Calculates  the hyperparameter lambda such that if sigma~Exp(lambda) then sigma>sigma0 with probability alpha
}
\examples{
alpha_sigma <- 0.01
sigma0 <- 10
lambda_sigma <- lambda_variance_quantile(alpha_sigma = alpha_sigma, sigma0 = sigma0)
}
