% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{log_gaussian_density}
\alias{log_gaussian_density}
\title{Log Gaussian density 1D}
\usage{
log_gaussian_density(x, mu, log_sigma)
}
\arguments{
\item{x}{A numeric vector representing the point x}

\item{mu}{A numeric vector representing the mean mu}

\item{log_sigma}{A numeric vector representing the variance log sigma}
}
\value{
The calculated log Gaussian density
}
\description{
Calculates  the log of the Gaussian density with mean mu and variance sigma^2
}
\examples{
x <- 1
mu <- 0
log_sigma <- 1
log_gaussian_density(x, mu, log_sigma)
}
