% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{log_gaussian_density}
\alias{log_gaussian_density}
\title{Log gaussian 1D density}
\usage{
log_gaussian_density(x, mu, logsigma)
}
\arguments{
\item{x}{A numeric vector representing the point x}

\item{mu}{A numeric vector representing the mean mu}

\item{logsigma}{A numeric vector representing the variance logsigma}
}
\value{
The calculated log gaussian density
}
\description{
Calculates  the log of the gaussian density with mean mu and variance sigma^2
}
\examples{
x <- 1
mu <- 0
logsigma <- 1
log_gaussian_density(x, mu, logsigma)
}
