% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{parameter_within_credible_intervals}
\alias{parameter_within_credible_intervals}
\title{Parameter within credible intervals}
\usage{
parameter_within_credible_intervals(parameter, credible_intervals)
}
\arguments{
\item{parameter}{A vector of length d representing the parameter}

\item{credible_intervals}{A matrix of size d x 2 representing the credible intervals}
}
\value{
A vector of length d representing whether the parameter is within the credible intervals
}
\description{
Checks if the parameter is within the credible intervals
}
\examples{
mu <- c(0, 0)
standard_deviation <- c(1, 1)
q <- 0.05
credible_intervals <- calculate_credible_intervals_gaussian(mu, standard_deviation, q)
theta <- c(0.5, 0.5)
parameter_within_credible_intervals(theta, credible_intervals)

}
